Refactor transcription and translation services to use a unified model manager
- Implemented intelligent model routing in `transcription_service.py` to choose between fast and accurate models based on audio characteristics and system load.
- Removed the old `transcription_service_faster_whisper.py` and `translation.py` files as their functionality has been integrated into the new structure.
- Enhanced the `translation_engine.py` to load models from the new `model_manager.py`, which centralizes model management and configuration.
- Added comprehensive logging for model loading, warmup, and usage statistics.
- Introduced a new `model_core.py` to handle the core components of the faster-whisper models, including loading, warmup, and statistics management.
- Updated `translation_utils.py` to include translation functions directly, eliminating the need for a separate translation module.
- Established a global instance of the `FasterWhisperModelManager` for consistent model management across the application.
