+++ code_stats_history.json
@@ -440,5 +440,70 @@
     "total_classes": 70,
     "total_imports": 234,
     "total_size_bytes": 326260
+  },
+  {
+    "analysis_date": "2025-07-06T21:56:37.292975",
+    "total_files": 30,
+    "total_lines": 7649,
+    "total_code_lines": 4938,
+    "total_comment_lines": 569,
+    "total_docstring_lines": 720,
+    "total_blank_lines": 1422,
+    "total_functions": 210,
+    "total_classes": 70,
+    "total_imports": 209,
+    "total_size_bytes": 316381
+  },
+  {
+    "analysis_date": "2025-07-06T21:57:21.884834",
+    "total_files": 30,
+    "total_lines": 7649,
+    "total_code_lines": 4938,
+    "total_comment_lines": 569,
+    "total_docstring_lines": 720,
+    "total_blank_lines": 1422,
+    "total_functions": 210,
+    "total_classes": 70,
+    "total_imports": 209,
+    "total_size_bytes": 316381
+  },
+  {
+    "analysis_date": "2025-07-06T22:04:02.332308",
+    "total_files": 29,
+    "total_lines": 7188,
+    "total_code_lines": 4615,
+    "total_comment_lines": 544,
+    "total_docstring_lines": 693,
+    "total_blank_lines": 1336,
+    "total_functions": 196,
+    "total_classes": 63,
+    "total_imports": 196,
+    "total_size_bytes": 294810
+  },
+  {
+    "analysis_date": "2025-07-06T22:12:04.838091",
+    "total_files": 29,
+    "total_lines": 7188,
+    "total_code_lines": 4615,
+    "total_comment_lines": 544,
+    "total_docstring_lines": 693,
+    "total_blank_lines": 1336,
+    "total_functions": 196,
+    "total_classes": 63,
+    "total_imports": 196,
+    "total_size_bytes": 294810
+  },
+  {
+    "analysis_date": "2025-07-06T22:22:49.973304",
+    "total_files": 29,
+    "total_lines": 7202,
+    "total_code_lines": 4626,
+    "total_comment_lines": 544,
+    "total_docstring_lines": 693,
+    "total_blank_lines": 1339,
+    "total_functions": 196,
+    "total_classes": 63,
+    "total_imports": 197,
+    "total_size_bytes": 295355
   }
 ]
