+++ config.yaml
@@ -40,3 +40,10 @@ audio:
   
 discord:
   # Discord bot configuration will go here
+
+translation:
+  # DeepL API configuration
+  deepl_api_key: "5ac935ea-9ed2-40a7-bd4d-8153c941c79f:fx"  # Replace with your DeepL API key
+  deepl_api_url: "https://api-free.deepl.com/v2/translate"   # Use https://api.deepl.com/v2/translate for paid accounts
+  target_language: "EN"  # Target language for translation
+  timeout: 10  # Request timeout in seconds
