+++ model_core.py
