--- requirements-faster-whisper.txt
@@ -1,34 +1,0 @@
-# Core Discord and voice processing - temporary PR branch fix for voice protocol v8 issues
-# This is a temporary fix for Discord's voice protocol v8 upgrade (WebSocket 4006 errors)
-# Install from DA-344's PR branch that contains the voice protocol v8 fix
-# TODO: Revert to official py-cord release once PR #2812 is merged and released
-git+https://github.com/DA-344/pycord.git@fix/voice-issues
-PyNaCl>=1.4.0
-
-# Machine learning and transcription - FASTER WHISPER IMPLEMENTATION (WORKING VERSIONS!)
-numpy<2.0  # Required for whisper model compatibility
-# IMPORTANT: These exact versions are WORKING with CUDA 12.1 + cuDNN 9.1 - DO NOT CHANGE!
---extra-index-url https://download.pytorch.org/whl/cu121
-torch==2.5.1+cu121      # WORKING: CUDA 12.1 with cuDNN 9.1.0 (90100)
-torchaudio==2.5.1+cu121  # WORKING: Compatible with above torch version
-faster-whisper==1.1.1    # WORKING: Latest version with bug fixes
-ctranslate2==4.5.0       # WORKING: Compatible with cuDNN 9.1 (downgraded from 4.6.0)
-soundfile>=0.13.1        # For audio file processing
-av>=11                   # Audio/video processing
-tokenizers>=0.13,<1      # Text tokenization
-huggingface-hub>=0.13    # Model downloads
-onnxruntime>=1.14,<2     # ONNX runtime
-
-# Web framework and API
-fastapi>=0.115.12
-uvicorn[standard]
-websockets>=15.0.1
-
-# System monitoring and utilities
-psutil>=7.0.0            # System process monitoring (required by server.py)
-
-# Configuration management
-PyYAML==6.0.1
-
-# HTTP requests for DeepL API
-requests==2.32.3
