+++ translation_utils.py
@@ -2,18 +2,93 @@
 Audio transcription and translation utilities.
 
 This module provides the main interface for transcription and translation
-functionality, now using the unified ModelManager for better organization.
+functionality, along with DeepL translation services and optimization logic.
 """
+import requests
 from logging_config import get_logger
+from config_manager import get_config
 
 # Import the faster-whisper model manager
-from faster_whisper_manager import faster_whisper_model_manager
+from model_manager import faster_whisper_model_manager
 import transcription_service
-import translation
 
 logger = get_logger(__name__)
 
-# Re-export the main functions for backward compatibility
+# Translation functions (moved from translation.py)
+
+
+async def translate(text):
+    """Translate text to English using DeepL's API.
+
+    Sends the provided text to DeepL's translation API and returns
+    the English translation. Configuration is loaded from config.yaml.
+
+    Args:
+        text (str): The text to be translated
+
+    Returns:
+        str: The translated text in English
+    """
+    config = get_config()
+    translation_config = config.get('translation', {})
+
+    api_key = translation_config.get('deepl_api_key')
+    api_url = translation_config.get(
+        'deepl_api_url', 'https://api-free.deepl.com/v2/translate')
+    target_lang = translation_config.get('target_language', 'EN')
+    timeout = translation_config.get('timeout', 10)
+
+    if not api_key:
+        logger.error("DeepL API key not configured in config.yaml")
+        raise ValueError("DeepL API key not configured")
+
+    response = requests.post(
+        api_url,
+        data={
+            "auth_key": api_key,
+            "text": text,
+            "target_lang": target_lang
+        },
+        timeout=timeout
+    )
+    translation = response.json()["translations"][0]["text"]
+    return translation
+
+
+async def should_translate(text, detected_language):
+    """Determine if text needs translation based on language and content.
+
+    This function optimizes translation API usage by avoiding unnecessary
+    translations of English text, while still detecting mixed-language content
+    that might need translation.
+
+    Args:
+        text (str): The text to potentially translate
+        detected_language (str): Language code detected by Whisper
+
+    Returns:
+        bool: True if text should be translated, False otherwise
+    """
+    # Always skip empty text
+    if not text:
+        return False
+
+    # Skip if dominant language is English
+    if detected_language == "en":
+        # Check if there are likely non-English segments
+        # This simple heuristic checks for characters common in non-Latin alphabets
+        non_ascii_ratio = len([c for c in text if ord(c) > 127]) / len(text)
+        if non_ascii_ratio > 0.1:  # If more than 10% non-ASCII, translate anyway
+            logger.info(
+                "Detected mixed language content (non-ASCII ratio: %.2f - translating)",
+                non_ascii_ratio)
+            return True
+        return False
+
+    # Non-English dominant language should be translated
+    return True
+
+# Transcription utility functions
 
 
 async def transcribe(audio_file, current_queue_size=0, concurrent_requests=0, active_transcriptions=0,
@@ -44,9 +119,6 @@ async def transcribe_with_timestamp(audio_file, current_queue_size=0, concurrent
         audio_file, current_queue_size, concurrent_requests, active_transcriptions,
         force_language, audio_timestamp
     )
-
-translate = translation.translate
-should_translate = translation.should_translate
 
 
 async def warm_up_pipeline():
